requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pytz==2024.1
orjson==3.10.3
httpx==0.27.0
python-dotenv==1.0.1

# NLP / ML
numpy==1.26.4
faiss-cpu==1.8.0.post1
fastembed==0.3.3
onnxruntime==1.18.1
langdetect==1.0.9
sentence-transformers>=2.2.2

# Integrations
twilio==9.0.5
openai>=1.30.0

# Document parsing / scraping
python-docx==0.8.11
pdfminer.six==20240706
beautifulsoup4==4.12.3

# Utils
fastapi-utils==0.7.0
typing-inspect